backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.31.0
cffi==2.0.0
click==8.3.1
cross-web==0.4.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
graphql-core==3.2.7
greenlet==3.3.0
h11==0.16.0
idna==3.11
packaging==25.0
passlib==1.7.4
psycopg2==2.9.11
pwdlib==0.3.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
sqlmodel==0.0.31
starlette==0.50.0
strawberry-graphql==0.288.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
alembic
